+++ app/src/main/AndroidManifest.xml
@@ -10,10 +10,10 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.ClockPackage">
         <activity
-            android:name=".ui.alarm.edit.AlarmEditActivity"
-            android:label="@string/title_activity_alarm_detail"></activity>
+            android:name="com.quokkaman.android.app.ui.edit.AlarmEditActivity"
+            android:label="@string/title_activity_alarm_detail" />
         <activity
-            android:name=".MainActivity"
+            android:name=".ui.main.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
