+++ app/src/main/java/com/quokkaman/android/app/ui/main/alarm/AlarmFragment.kt
@@ -1,4 +1,4 @@
-package com.quokkaman.android.app.ui.alarm
+package com.quokkaman.android.app.ui.main.alarm
 
 import androidx.lifecycle.ViewModelProvider
 import android.os.Bundle
