+++ app/src/main/res/layout/item_edit_alarm_plan.xml
@@ -18,7 +18,7 @@
 
         <variable
             name="viewModel"
-            type="com.quokkaman.android.app.ui.alarm.edit.AlarmPlanViewModel" />
+            type="com.quokkaman.android.app.ui.edit.AlarmPlanViewModel" />
 
     </data>
