+++ app/src/main/res/layout/stop_watch_fragment.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.stopwatch.StopWatchFragment">
+    tools:context=".ui.main.stopwatch.StopWatchFragment">
 
     <TextView
         android:layout_width="match_parent"
