+++ app/src/main/res/layout/timer_fragment.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.timer.TimerFragment">
+    tools:context=".ui.main.timer.TimerFragment">
 
     <TextView
         android:layout_width="match_parent"
