+++ app/src/main/res/layout/world_clock_fragment.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.world.WorldClockFragment">
+    tools:context=".ui.main.world.WorldClockFragment">
 
     <TextView
         android:layout_width="match_parent"
